refactor: separate files into different folders
