+++ oxide_math/lib.rs
@@ -3,7 +3,3 @@ pub mod matrix;
 pub mod matrix2x2;
 pub mod quaternion;
 pub mod test;
-pub mod vector;
-pub mod vector2;
-pub mod vector3;
-pub mod vector4;
