+++ oxide_math/test/vector2.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod test {
     use crate::test::utils::*;
-    use crate::vector::*;
-    use crate::{vector2, vector3};
+    use crate::commons::vector::*;
+    use crate::commons::{vector2, vector3};
 
     #[test]
     fn get_squared_length() {
