+++ oxide_math/test/vector4.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod test {
     use crate::test::utils::*;
-    use crate::vector::*;
-    use crate::vector4;
+    use crate::commons::vector::*;
+    use crate::commons::vector4;
 
     #[test]
     fn get_squared_length() {
